style: reformat functions.py
